Prevent buffer overflows
Modify functions that take a `u_char *buf` and return an updated pointer, to
instead take a `ngx_buf_t`. This allows use of `slprintf` to prevent buffer
overflows.  We don't yet check for truncation if the buffer is filled, but
modified the functions to return a status code to allow this.